:bug: fix:기존 잘못 만들어진 경로 수정 goal -> goal/id
