+++ src/app/(routers)/(todo)/goals/[goalId]/todos/[todoId]/page.tsx
@@ -1,0 +1,3 @@
+export default async function TodoIdPage() {
+  return <div>todoIdPage</div>;
+}
